+++ src/headers/utils.h
@@ -12,7 +12,7 @@ Data* findSubObjectBySCIndex(Data* object, uint64_t index);
 Data* findObjectByHeaderAddress(address_t address);
 uint16_t getRealNameLength(Data* object);
 void parseCoordinates(VariableNameToken* vnt);
-uint64_t coordToInd(const uint32_t* coords, const uint32_t* dims, uint8_t num_dims);
+uint64_t coordToInd(const uint64_t* coords, const uint64_t* dims, uint8_t num_dims);
 void makeVarnameQueue(char* variable_name);
 Data* cloneData(Data* old_object);
 void removeSpaces(char* source);
